fixed juice device tiles
Why they were working before, I don't know.  Caused issues for drawers, and might cause issues for other modules that get confused by the non-standard tiles field.

See:
https://minetest.gitlab.io/minetest/definition-tables/#node-definition
https://minetest.gitlab.io/minetest/definition-tables/#tile-definition

A tile definition may be a table, but one or more tile definitions must go _inside_ a list as part of a node definition.  I guess the engine might support other configurations (it was working), but if so, it's not documented as safe practice.